Rebase conflict fixes
