+++ cmd/webhook-client/post_webhook_notify.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"encoding/json"
+	"errors"
 	"fmt"
 
 	"github.com/spf13/cobra"
@@ -23,18 +24,18 @@ type WebhookRequest struct {
 
 // Flags specific to this command
 const (
-	// MessageFlag is the string to send in the payload
-	MessageFlag string = "message"
+	// FilenameFlag is the string to send in the payload
+	FilenameFlag string = "filename"
 )
 
 func initPostWebhookNotifyFlags(flag *pflag.FlagSet) {
-	flag.String(MessageFlag, "Hello World", "Message for the client to send")
+	flag.String(FilenameFlag, "", "Filename of json file to send")
 	flag.SortFlags = false
 }
 
 func checkPostWebhookNotifyConfig(v *viper.Viper, args []string, logger utils.Logger) error {
 
-	missingFilenameFlag := v.GetString(FilenameFlag) == "" && (len(args) < 1 || len(args) > 0 && !ContainsDash(args))
+	missingFilenameFlag := v.GetString(FilenameFlag) == "" && (len(args) < 1 || len(args) > 0 && !utils.ContainsDash(args))
 
 	if missingFilenameFlag {
 		logger.Fatal("post-webhook-notify expects --filename with json file passed in")
@@ -68,7 +69,7 @@ func postWebhookNotify(cmd *cobra.Command, args []string) error {
 	// Decode the json file that was passed in
 	filename := v.GetString(FilenameFlag)
 	payload := &WebhookRequest{}
-	err = DecodeJSONFileToPayload(filename, ContainsDash(args), &payload)
+	err = utils.DecodeJSONFileToPayload(filename, utils.ContainsDash(args), &payload)
 
 	if err != nil {
 		logger.Error("Error opening file:", zap.Error(err))
@@ -101,7 +102,6 @@ func postWebhookNotify(cmd *cobra.Command, args []string) error {
 	path := "support/v1/webhook-notify"
 
 	url := fmt.Sprintf("https://%s:%d/%s", hostname, port, path)
-	fmt.Println("url is", url)
 	resp, body, err := runtime.Post(json, url)
 
 	if err != nil {
