+++ cmd/webhook-client/utils/shared.go
@@ -1,7 +1,12 @@
 package utils
 
 import (
+	"bufio"
+	"encoding/json"
+	"errors"
 	"fmt"
+	"os"
+	"path/filepath"
 	"strings"
 
 	"github.com/spf13/cobra"
@@ -47,4 +52,45 @@ func ParseFlags(cmd *cobra.Command, v *viper.Viper, args []string) error {
 	v.SetEnvKeyReplacer(strings.NewReplacer("-", "_"))
 	v.AutomaticEnv()
 	return nil
+}
+
+// DecodeJSONFileToPayload takes a filename, or stdin and decodes the file into
+// the supplied json payload.
+// If the filename is not supplied, the isStdin bool should be set to true to use stdin.
+// If the file contains parameters that do not exist in the payload struct, it will fail with an error
+// Otherwise it will populate the payload
+func DecodeJSONFileToPayload(filename string, isStdin bool, payload interface{}) error {
+	var reader *bufio.Reader
+	if filename != "" {
+		file, err := os.Open(filepath.Clean(filename))
+		if err != nil {
+			return fmt.Errorf("File open failed: %w", err)
+		}
+		reader = bufio.NewReader(file)
+	} else if isStdin { // Uses std in if "-"" is provided instead
+		reader = bufio.NewReader(os.Stdin)
+	} else {
+		return errors.New("no file input was found")
+	}
+
+	jsonDecoder := json.NewDecoder(reader)
+	jsonDecoder.DisallowUnknownFields()
+
+	// Read the json into the mto payload
+	err := jsonDecoder.Decode(payload)
+	if err != nil {
+		return fmt.Errorf("File decode failed: %w", err)
+	}
+
+	return nil
+}
+
+// ContainsDash returns true if the original command included an empty dash
+func ContainsDash(args []string) bool {
+	for _, arg := range args {
+		if arg == "-" {
+			return true
+		}
+	}
+	return false
 }
